pyproject.toml
[project]
name = "vlm-sam-industrial-vision-v2"
version = "0.1.0"
description = "RK3588 edge-to-cloud industrial vision defect detection system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.9",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "websockets>=13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
